requirements.txt
requests
dogpile.cache
